// packages/3-extensions/sql-orm-client/test/fixtures/fragment-namespace/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly fragment: {
    readonly Audit: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly message: CodecTypes['pg/text@1']['output'];
    };
  };
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly fragment: {
    readonly Audit: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly message: CodecTypes['pg/text@1']['input'];
    };
  };
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly fragment: {
    readonly audits: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly message: CodecTypes['pg/text@1']['output'];
    };
  };
  readonly public: {
    readonly posts: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly fragment: {
    readonly audits: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly message: CodecTypes['pg/text@1']['input'];
    };
  };
  readonly public: {
    readonly posts: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;